+++ packages/create-rainbowkit/templates/next-app/pages/_app.tsx
@@ -9,6 +9,7 @@ import {
   mainnet,
   optimism,
   polygon,
+  base,
   zora,
 } from 'wagmi/chains';
 import { publicProvider } from 'wagmi/providers/public';
@@ -19,6 +20,7 @@ const { chains, publicClient, webSocketPublicClient } = configureChains(
     polygon,
     optimism,
     arbitrum,
+    base,
     zora,
     ...(process.env.NEXT_PUBLIC_ENABLE_TESTNETS === 'true' ? [goerli] : []),
   ],