+++ packages/rainbowkit/src/wallets/walletConnectors/index.ts
@@ -1,4 +1,5 @@
 import { argentWallet } from './argentWallet/argentWallet';
+import { bifrostWallet } from './bifrostWallet/bifrostWallet';
 import { bitKeepWallet } from './bitKeepWallet/bitKeepWallet';
 import { bitskiWallet } from './bitskiWallet/bitskiWallet';
 import { braveWallet } from './braveWallet/braveWallet';
@@ -6,6 +7,8 @@ import { coinbaseWallet } from './coinbaseWallet/coinbaseWallet';
 import { coreWallet } from './coreWallet/coreWallet';
 import { dawnWallet } from './dawnWallet/dawnWallet';
 import { enkryptWallet } from './enkryptWallet/enkryptWallet';
+import { foxWallet } from './foxWallet/foxWallet';
+import { frameWallet } from './frameWallet/frameWallet';
 import { frontierWallet } from './frontierWallet/frontierWallet';
 import { imTokenWallet } from './imTokenWallet/imTokenWallet';
 import { injectedWallet } from './injectedWallet/injectedWallet';
@@ -23,12 +26,14 @@ import { tahoWallet } from './tahoWallet/tahoWallet';
 import { talismanWallet } from './talismanWallet/talismanWallet';
 import { tokenPocketWallet } from './tokenPocketWallet/tokenPocketWallet';
 import { trustWallet } from './trustWallet/trustWallet';
+import { uniswapWallet } from './uniswapWallet/uniswapWallet';
 import { walletConnectWallet } from './walletConnectWallet/walletConnectWallet';
 import { xdefiWallet } from './xdefiWallet/xdefiWallet';
 import { zerionWallet } from './zerionWallet/zerionWallet';
 
 export {
   argentWallet,
+  bifrostWallet,
   bitKeepWallet,
   bitskiWallet,
   braveWallet,
@@ -36,12 +41,15 @@ export {
   coreWallet,
   dawnWallet,
   enkryptWallet,
+  foxWallet,
+  frameWallet,
   frontierWallet,
   imTokenWallet,
   injectedWallet,
   ledgerWallet,
   metaMaskWallet,
   mewWallet,
+  okxWallet,
   omniWallet,
   phantomWallet,
   rabbyWallet,
@@ -52,7 +60,7 @@ export {
   talismanWallet,
   tokenPocketWallet,
   trustWallet,
-  okxWallet,
+  uniswapWallet,
   walletConnectWallet,
   xdefiWallet,
   zerionWallet,