+++ site/components/Provider/Provider.tsx
@@ -11,12 +11,20 @@ import {
 } from '@rainbow-me/rainbowkit/wallets';
 import React from 'react';
 import { configureChains, createConfig, WagmiConfig } from 'wagmi';
-import { arbitrum, bsc, mainnet, optimism, polygon, zora } from 'wagmi/chains';
+import {
+  arbitrum,
+  base,
+  bsc,
+  mainnet,
+  optimism,
+  polygon,
+  zora,
+} from 'wagmi/chains';
 import { alchemyProvider } from 'wagmi/providers/alchemy';
 import { publicProvider } from 'wagmi/providers/public';
 
 export const { chains, publicClient } = configureChains(
-  [mainnet, polygon, optimism, arbitrum, bsc, zora],
+  [mainnet, polygon, optimism, arbitrum, base, zora, bsc],
   [
     alchemyProvider({ apiKey: process.env.NEXT_PUBLIC_ALCHEMY_ID ?? '' }),
     publicProvider(),