+++ site/CHANGELOG.md
@@ -1,5 +1,12 @@
 # site
 
+## 0.0.70
+
+### Patch Changes
+
+- Updated dependencies [eb319f3]
+  - @rainbow-me/rainbowkit@1.0.8
+
 ## 0.0.69
 
 ### Patch Changes